refactor: move useDynamic{Route,Search}Params to reduce snapshot churn
